refactor: merge station related endpoints into single file
