+++ alerts/urls.py
@@ -2,9 +2,12 @@ from django.urls import path
 
 from .views.map import get_map
 
-from .views import formula, report, view_graphs, collect_data_cepadi, get_sensor_data, \
-    collect_data_stationif, \
-    get_station_sensors_data, download_data_station, get_station, get_mathmodels, get_station_mathmodel_color, webhooks
+from .views import formula, report, view_graphs, \
+    collect_data_cepadi, collect_data_stationif, download_data_station, \
+    get_sensor_data, \
+    get_station, get_station_sensors_data, get_station_mathmodel_color, \
+    get_mathmodels, \
+    webhooks
 from .views import mathmodel
 
 app_name = 'alerts'
@@ -22,11 +25,16 @@ urlpatterns = [
     path('coletar_dados/cepadi', collect_data_cepadi, name="collect_data_cepadi"),
     path('coletar_dados', report.ReportView.as_view(), name="collect_data"),
     path('get/sensor/<int:sensor_id>/<str:date_filter>/data', get_sensor_data, name="get_sensor_data"),
-    path('get/estacao/<str:station_id>/sensores', get_station_sensors_data, name="get_station_data"),
-    path('baixar/dados/estacao/<str:station_id>', download_data_station, name="dowload_data_station"),
-    path('get/estacao/<str:station_id>', get_station, name="get_station"),
+
     path('get/mathmodels/<str:date_filter>', get_mathmodels, name="get_mathmodels"),
+    path('baixar/dados/estacao/<str:station_id>', download_data_station, name="dowload_data_station"),
+
+    # Station
+    path('get/station/<str:station_id>/sensores', get_station_sensors_data, name="get_station_data"),
+    path('get/station/<str:station_id>', get_station, name="get_station"),
     path('get/station/<str:station_id>/mathmodel/<int:mathmodel_id>/color', get_station_mathmodel_color,
          name="get_station_mathmodel_color"),
+
+    # Webhook
     path('webhooks/whatsapp', webhooks.whatsapp, name="whatsapp")
 ]
