+++ alerts/views/__init__.py
@@ -1,12 +1,10 @@
 from .report import *
-from .test import *
 from .mathmodel import *
 from .graph import *
 from .collect_data_cepadi import *
 from .get_sensor_data import *
 from .collect_data_stationif import *
-from .get_station_sensors_data import *
+from .station import *
 from .download_data_station import *
-from .get_station import *
 from .get_mathmodels import *
-from .get_station_mathmodel_color import *
+from .webhooks import *
